Clarified naming, types and got rid of platform-dependent integer sizes
Also, fixed the WASM build for morpheus

+++ coeus-core/src/lib.rs
@@ -35,9 +35,10 @@ use iop_keyvault::{
 #[cfg(feature = "did")]
 use iop_morpheus_core::data::did::Did;
 
+// TODO move all blockchain-related types to hydra-proto after adding typetags to Asset and TransactionType.
 /// State identifier of a decentralized ledger, usually represented as a sequence number of blocks.
-pub type BlockHeight = u64;
+pub type BlockHeight = u32;
 /// Duration (e.g. a year) expressed as an expected number of blocks on the ledger, approximating the duration.  
-pub type BlockCount = u64;
+pub type BlockCount = u32;
 /// A unique value attached to requests as protection from replay attacks.
 pub type Nonce = u64;
